Create Problem 3
This problem entails allowing the user to enter the number of airlines tickets and cost per seat for a purchase. (We will assume all seats are the same price). You are to use a function that receives the number of tickets and price per seat and calculates the cost, tax (7% of cost) and total (cost + tax).